Changes to how sections are recorded in the schedule. SectionTracker takes care of this.
Algorithm now only takes list of courses. If you want to set the instructor/locations for a schedule, call the appropriate constructor or use the setters. 

Changes availability to take a boolean value of whether you want to set as free or busy. Instructor and Location objects have been updated to reflect this new form of functionality.

git-svn-id: svn+ssh://scheduler.csc.calpoly.edu/repo@744 13b268f7-8392-4c9d-bc7d-1e80a86bc2f9